:fix: remove unnecessary Normal VisualState definition.
Because of the Normal VisualState definition, visual of PathViewItem change to explicitly defined state, which is not the visual user specified. As a result of removing this definition, the Normal visual state has no library defined visual and the visual of control turn back to the initial. But a new issue is appeared by this solution. When a PathViewItem is expanded and mouse pointer is left from it, visual of the PathViewItem don't translate to 'Normal' state even when its visual state is 'Normal'. This new issue will be fixed in the next commit.