:fix: fix #3
When a PathViewItem is expanded and mouse pointer is left from it, the visual state translate to the 'Normal' state but the 'Expand' state previously set does not stop in any reason. This caused a situaltion that the visual of PathViewItem did not back to 'only Normal' state and became 'Normal and Expand' state. The true reason of this phenomenon is not appeared. But the phenomenon can be avoided by explicitly defining 'Normal' state with empty 'To' satte.